+++ src/test/java/com/imsweb/layout/LayoutUtilsTest.java
@@ -53,7 +53,7 @@ public class LayoutUtilsTest {
     public void testInputOutputStreamCreation() throws IOException {
 
         // write to file
-        File file = new File(System.getProperty("user.dir") + "/build/create-input-stream-test.txt");
+        File file = new File(TestingUtils.getWorkingDirectory()+ "/build/create-input-stream-test.txt");
         try (OutputStream os = LayoutUtils.createOutputStream(file)) {
             Assert.assertNotNull(os);
             os.write("A".getBytes());
@@ -67,7 +67,7 @@ public class LayoutUtilsTest {
         }
 
         // write to gzipped file
-        File gzippedFile = new File(System.getProperty("user.dir") + "/build/create-input-stream-test.txt.gz");
+        File gzippedFile = new File(TestingUtils.getWorkingDirectory() + "/build/create-input-stream-test.txt.gz");
         try (OutputStream os = LayoutUtils.createOutputStream(gzippedFile)) {
             Assert.assertNotNull(os);
             os.write("A".getBytes());
@@ -81,7 +81,7 @@ public class LayoutUtilsTest {
         }
 
         // write to zipped file
-        File zippedFile = new File(System.getProperty("user.dir") + "/build/create-input-stream-test.zip");
+        File zippedFile = new File(TestingUtils.getWorkingDirectory() + "/build/create-input-stream-test.zip");
         try (OutputStream os = LayoutUtils.createOutputStream(zippedFile)) {
             Assert.assertNotNull(os);
             ((ZipOutputStream)os).putNextEntry(new ZipEntry("create-input-stream-test"));
@@ -118,7 +118,7 @@ public class LayoutUtilsTest {
         field.setSection("section");
         layout.setField(Collections.singletonList(field));
 
-        File file = new File(System.getProperty("user.dir") + "/build/fixed-layout-test.xml");
+        File file = new File(TestingUtils.getWorkingDirectory() + "/build/fixed-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeFixedColumnsLayout(fos, layout);
         }
@@ -164,7 +164,7 @@ public class LayoutUtilsTest {
         field.setDefaultValue("default");
         layout.setField(Collections.singletonList(field));
 
-        File file = new File(System.getProperty("user.dir") + "/build/fixed-layout-test.xml");
+        File file = new File(TestingUtils.getWorkingDirectory()+ "/build/fixed-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeCommaSeparatedLayout(fos, layout);
         }
@@ -219,7 +219,7 @@ public class LayoutUtilsTest {
         segment.setHl7Fields(Collections.singletonList(field));
         layout.setHl7Segments(Collections.singletonList(segment));
 
-        File file = new File(System.getProperty("user.dir") + "/build/hl7-layout-test.xml");
+        File file = new File(TestingUtils.getWorkingDirectory() + "/build/hl7-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeHl7Layout(fos, layout);
         }
@@ -283,7 +283,7 @@ public class LayoutUtilsTest {
         commaSeparatedField.setIndex(1);
         commaSeparatedLayout.setField(Collections.singletonList(commaSeparatedField));
 
-        File file = new File(System.getProperty("user.dir") + "/build/fixed-layout-test.xml");
+        File file = new File(TestingUtils.getWorkingDirectory() + "/build/fixed-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeCommaSeparatedLayout(fos, commaSeparatedLayout);
         }
@@ -303,7 +303,7 @@ public class LayoutUtilsTest {
         fixedColumnField.setEnd(1);
         fixedColumnLayout.setField(Collections.singletonList(fixedColumnField));
 
-        file = new File(System.getProperty("user.dir") + "/build/fixed-layout-test.xml");
+        file = new File(TestingUtils.getWorkingDirectory() + "/build/fixed-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeFixedColumnsLayout(fos, fixedColumnLayout);
         }
@@ -324,7 +324,7 @@ public class LayoutUtilsTest {
         segment.setHl7Fields(Collections.singletonList(field));
         layout.setHl7Segments(Collections.singletonList(segment));
 
-        file = new File(System.getProperty("user.dir") + "/build/hl7-layout-test.xml");
+        file = new File(TestingUtils.getWorkingDirectory() + "/build/hl7-layout-test.xml");
         try (OutputStream fos = new FileOutputStream(file)) {
             LayoutUtils.writeHl7Layout(fos, layout);
         }