+++ src/test/java/com/imsweb/layout/record/fixed/naaccr/Naaccr16LayoutTest.java
@@ -16,6 +16,7 @@ import org.junit.Test;
 
 import com.imsweb.layout.Field;
 import com.imsweb.layout.LayoutFactory;
+import com.imsweb.layout.TestingUtils;
 import com.imsweb.layout.record.fixed.FixedColumnsField;
 import com.imsweb.layout.record.fixed.FixedColumnsLayout;
 
@@ -69,7 +70,7 @@ public class Naaccr16LayoutTest {
         Assert.assertNull(rec.get("reserved01"));
 
         // test write methods        
-        File file = new File(System.getProperty("user.dir") + "/build/naaccr16.txt");
+        File file = new File(TestingUtils.getWorkingDirectory()+ "/build/naaccr16.txt");
         rec.clear();
         rec.put("naaccrRecordVersion", "160");
         rec.put("recordType", "I");