+++ src/test/java/com/imsweb/layout/record/fixed/naaccr/Naaccr18LayoutTest.java
@@ -17,6 +17,7 @@ import org.junit.Test;
 
 import com.imsweb.layout.Field;
 import com.imsweb.layout.LayoutFactory;
+import com.imsweb.layout.TestingUtils;
 import com.imsweb.layout.record.fixed.FixedColumnsField;
 import com.imsweb.layout.record.fixed.FixedColumnsLayout;
 
@@ -24,7 +25,7 @@ public class Naaccr18LayoutTest {
 
     @BeforeClass
     public static void setup() {
-        File buildFolder = new File(System.getProperty("user.dir") + "/build/");
+        File buildFolder = new File(TestingUtils.getWorkingDirectory() + "/build/");
         if (!buildFolder.exists() && !buildFolder.mkdir())
             Assert.fail("Can't create build folder");
     }
@@ -77,7 +78,7 @@ public class Naaccr18LayoutTest {
         Assert.assertNull(rec.get("reserved04"));
 
         // test write methods
-        File file = new File(System.getProperty("user.dir") + "/build/naaccr18.txt");
+        File file = new File(TestingUtils.getWorkingDirectory() + "/build/naaccr18.txt");
         rec.clear();
         rec.put("naaccrRecordVersion", "180");
         rec.put("recordType", "I");