+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Changes in version 1.11**
 
+- Updated NAACCR XML library to version 4.10.
 - Updated NAACCR 18 field documentation.
 - Set default user dictionary URI in NaaccrXmlLayout.buildFileInfo()